Move quiz building logic to QuizBuilderContext
