+++ client/src/components/QuizQuestions/Mockquiz.tsx
@@ -1,5 +1,6 @@
-import { useEffect, useState } from "react";
 import { StyledButton } from "../buttons/AddButton";
+import { useQuizBuilder } from "../../context/QuizBuilderContext";
+import { useEffect } from "react";
 
 // THESE INTERFACES ARE FOR THE QUESTIONS THAT ARE BEING CREATED TO SEND TO THE API
 // THAT IS WHY THE QUESTIONTYPE DOESN'T HAVE A _ID
@@ -14,74 +15,11 @@ export interface AnswerType {
 }
 
 function MockQuiz() {
-  const [questions, setQuestions] = useState<QuestionType[]>([]);
+  const { questions, addQuestions, updateQuestion } = useQuizBuilder();
 
   useEffect(() => {
     console.log(questions);
   }, [questions]);
-
-  const addQuestions = () => {
-    setQuestions([
-      ...questions,
-      {
-        question: "",
-        answers: [
-          { _id: "A", answer: "", isCorrect: true },
-          { _id: "B", answer: "", isCorrect: false },
-          { _id: "C", answer: "", isCorrect: false },
-          { _id: "D", answer: "", isCorrect: false },
-        ],
-      },
-    ]);
-  };
-
-  const updateQuestion = (
-    index: number,
-    field: "question" | "isCorrect" | "answers",
-    value: string | boolean,
-    answerId?: string
-  ) => {
-    const updatedQuestions = questions.map((question, i) => {
-      if (i === index) {
-        switch (field) {
-          case "question":
-            // Update the question text
-            return { ...question, question: value as string };
-
-          case "answers":
-            // Update the text of a specific answer
-            return {
-              ...question,
-              answers: question.answers.map((answer) => {
-                if (answer._id === answerId) {
-                  return { ...answer, answer: value as string };
-                }
-                return answer;
-              }),
-            };
-
-          case "isCorrect":
-            // Toggle the correctness of a specific answer
-            return {
-              ...question,
-              answers: question.answers.map((answer) => {
-                if (answer._id === answerId) {
-                  return { ...answer, isCorrect: true };
-                } else {
-                  return { ...answer, isCorrect: false };
-                }
-              }),
-            };
-
-          default:
-            // In case an unknown field is used
-            return question;
-        }
-      }
-      return question;
-    });
-    setQuestions(updatedQuestions);
-  };
 
   const focusStyle =
     "focus:outline-none focus:ring-2 focus:ring-primary focus:ring-opacity-50";
