+++ client/src/pages/QuizOverviewPage.tsx
@@ -1,8 +1,13 @@
-import QuizOverview from "../components/quiz/QuizOverview";
+import MockQuiz from "../components/QuizQuestions/Mockquiz";
+import { QuizProvider } from "../context/QuizBuilderContext";
 import { useRequireAuth } from "../hooks/useRequireAuth";
 
 export default function QuizOverviewPage() {
   useRequireAuth();
 
-  return <QuizOverview />;
+  return (
+    <QuizProvider>
+      <MockQuiz />
+    </QuizProvider>
+  );
 }
